change 'long long' tom int64_t'; change variable buf to new delete;
git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@291 e54a6e83-5905-42d5-8d5c-058d10e6a962

+++ lib/libdvbsub/helpers.cpp
@@ -23,7 +23,7 @@ uint32_t getbits(const uint8_t* buf, uint32_t offset, uint8_t len)
 		len -= 8;
 	}
 	if (len > 8) {
-		unsigned long long tmp = retval << 8;
+		uint64_t tmp = retval << 8;
 		tmp |= *(a+4);
 		tmp >>= ((8-(offset%8)) + (8-(len)));
 		return tmp & (mask -1);
